median_filter becoming medianfilter and creating a new file MedianFilterDialog under silx.gui.widgets to deal with the median filter user interface
